pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-instructions-app"
version = "0.1.0"
description = "Web application for AI instructions generation"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "AI Team"}]
dependencies = [
  "Flask>=3.0",
  "flask-cors>=4.0",
  "Flask-SQLAlchemy>=3.1",
  "pydantic>=2.5",
  "langchain>=0.2",
  "langgraph>=0.1.0",
  "langchain-openai>=0.1.0",
  "redis>=5.0",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "gunicorn>=21.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "black>=24.0",
]
